refactor: keep static page chrome on the server
Move home, news, mails, portfolio, and chess shells to Server Components with client islands for state and events. Use useMounted, useOptimistic on mails, and useEffectEvent for chess save flush so client JS is limited to interactive UI.
