+++ src/app/api/photos/route.ts
@@ -1,10 +1,8 @@
 import { auth } from '@/auth'
 import { resolveCapturedAt } from '@/lib/photo-metadata.server'
 import { getPhotoDisplayDate } from '@/lib/photo-metadata'
-import {
-  resolveContentType,
-  validatePhotoFile
-} from '@/lib/photos'
+import { photoUploadFieldsSchema } from '@/lib/photo-upload-validation'
+import { resolveContentType, validatePhotoFile } from '@/lib/photos'
 import { uploadFileToR2 } from '@/lib/r2'
 import { Photo } from '@/models/Photo'
 import dbConnect from '@/utils/db'
@@ -50,18 +48,33 @@ export const POST = async (req: NextRequest) => {
     await dbConnect()
 
     const formData = await req.formData()
-    const file = formData.get('file') as File
-    const title = formData.get('title') as string
-    const description = formData.get('description') as string
-    const isPublic = formData.get('isPublic') === 'true'
-    const manualCapturedAt = formData.get('capturedAt') as string | null
+    const file = formData.get('file')
+    const titleField = formData.get('title')
+    const descriptionField = formData.get('description')
+    const manualCapturedAt = formData.get('capturedAt')
 
-    if (!file || !title) {
-      return NextResponse.json(
-        { error: 'File and title are required' },
-        { status: 400 }
-      )
+    if (!(file instanceof File)) {
+      return NextResponse.json({ error: 'File is required' }, { status: 400 })
     }
+
+    const validatedFields = photoUploadFieldsSchema.safeParse({
+      title: typeof titleField === 'string' ? titleField : '',
+      description: typeof descriptionField === 'string' ? descriptionField : '',
+      isPublic: formData.get('isPublic') === 'true',
+      capturedAt:
+        typeof manualCapturedAt === 'string' && manualCapturedAt.length > 0
+          ? manualCapturedAt
+          : undefined
+    })
+
+    if (!validatedFields.success) {
+      const message =
+        validatedFields.error.issues[0]?.message ?? 'Invalid upload metadata'
+      return NextResponse.json({ error: message }, { status: 400 })
+    }
+
+    const { title, description, isPublic, capturedAt: manualCapturedAtValue } =
+      validatedFields.data
 
     const validationError = validatePhotoFile(file)
     if (validationError) {
@@ -83,10 +96,10 @@ export const POST = async (req: NextRequest) => {
     // Upload to R2
     const imageUrl = await uploadFileToR2(buffer, key, contentType)
 
-    const capturedAt = await resolveCapturedAt(
+    const resolvedCapturedAt = await resolveCapturedAt(
       buffer,
       file.name,
-      manualCapturedAt
+      manualCapturedAtValue ?? null
     )
 
     // Save to database
@@ -97,7 +110,7 @@ export const POST = async (req: NextRequest) => {
       imageKey: key,
       uploadedBy: session.user.email,
       isPublic,
-      ...(capturedAt ? { capturedAt } : {})
+      ...(resolvedCapturedAt ? { capturedAt: resolvedCapturedAt } : {})
     })
 
     await photo.save()
