+++ src/app/chess/page.tsx
@@ -1,7 +1,14 @@
 import ChessLayout from '@/app/chess/_components/ChessLayout'
+import { Montserrat } from 'next/font/google'
 
-const ChessPage = () => {
-  return <ChessLayout />
-}
+const montserrat = Montserrat({ subsets: ['latin'], display: 'swap' })
+
+const ChessPage = () => (
+  <main
+    className={`flex overflow-hidden h-full w-full xl:max-w-300 sm:pt-6 xl:pt-12 lg:max-w-275 mx-auto sm:px-12 lg:px-0 ${montserrat.className} my-2 sm:my-0`}
+  >
+    <ChessLayout />
+  </main>
+)
 
 export default ChessPage
