+++ src/app/portfolio/page.tsx
@@ -1,489 +1,14 @@
-'use client'
-
-import Image from 'next/image'
-
-import { education, projects, skillCategories, skills, workExperiences } from '@/constants'
-import {
-  CodeBracketIcon,
-  SquaresPlusIcon,
-  BriefcaseIcon,
-  BookOpenIcon,
-  LinkIcon
-} from '@heroicons/react/24/outline'
-import { useState } from 'react'
+import Portfolio from '@/app/portfolio/_components/Portfolio'
 import { Montserrat } from 'next/font/google'
-import {
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  XMarkIcon
-} from '@heroicons/react/24/solid'
-import CompanyLogo from '@/app/portfolio/_components/CompanyLogo'
-import SkillLogo from '@/app/portfolio/_components/SkillLogo'
 
 const montserrat = Montserrat({ subsets: ['latin'], display: 'swap' })
 
-const Portfolio = () => {
-  const [activeCategory, setActiveCategory] = useState<number | null>(null)
+const PortfolioPage = () => (
+  <main
+    className={`flex overflow-hidden h-full w-full xl:max-w-[1024px] sm:pt-6 xl:pt-12 lg:max-w-[924px] mx-auto sm:px-12 lg:px-0 ${montserrat.className} my-2 sm:my-0`}
+  >
+    <Portfolio />
+  </main>
+)
 
-  const icons = [
-    {
-      title: 'Skills',
-      icon: CodeBracketIcon
-    },
-    {
-      title: 'Projects',
-      icon: SquaresPlusIcon
-    },
-    {
-      title: 'Work',
-      icon: BriefcaseIcon
-    },
-    {
-      title: 'Certificates',
-      icon: BookOpenIcon
-    }
-  ]
-
-  return (
-    <main
-      className={`flex overflow-hidden h-full w-full xl:max-w-[1024px] sm:pt-6 xl:pt-12 lg:max-w-[924px] mx-auto sm:px-12 lg:px-0 ${montserrat.className} my-2 sm:my-0`}
-    >
-      <div className='flex grow h-full rounded-xl bg-stone-200/95 dark:bg-zinc-900 flex-col shadow-xl overflow-hidden'>
-        <div className='flex border-b-[0.3px] border-gray-400/60 sm:border-gray-300/90 dark:border-zinc-700 border-solid'>
-          <div className='px-4 xs:pt-5 hidden xs:pb-3 xs:flex gap-x-4 items-center hover:bg-[#F5F5F5] dark:hover:bg-zinc-800 md:w-1/3 shadow-xs'>
-            <div className='block'>
-              <div className='relative h-[60px] w-[60px]'>
-                <Image
-                  sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-                  src={'/images/icons/portfolio-thumbnail.png'}
-                  fill
-                  priority
-                  alt='Portfolio Thumbnail'
-                />
-              </div>
-            </div>
-            <div className='flex flex-col'>
-              <h1 className='font-medium text-base text-black dark:text-white'>Portfolio</h1>
-              <p className='font-normal text-xs text-gray-700 dark:text-gray-300 tracking-tight'>
-                Get to know about Jerald here.
-              </p>
-            </div>
-          </div>
-          <div
-            className={`flex-1 px-4 pt-2.5 pb-3 hidden md:flex gap-x-4 h-[92px] flex-col justify-between items-center bg-white/95 dark:bg-zinc-800/95 overflow-hidden ${
-              activeCategory !== null ? 'shadow-sm' : ''
-            }`}
-          >
-            <div className='flex gap-1'>
-              {icons.map((_, i) => (
-                <button
-                  className={`h-1.5 w-1.5 rounded-full ${
-                    activeCategory === i ? 'bg-gray-500' : 'bg-gray-400'
-                  }`}
-                  key={_.title}
-                  onClick={() => setActiveCategory(i)}
-                />
-              ))}
-            </div>
-            {activeCategory != null && (
-              <div className='flex items-center justify-between h-11 w-full'>
-                <div className='flex gap-x-1 w-12 justify-between'>
-                  {activeCategory > 0 ? (
-                    <ChevronLeftIcon
-                      className={`text-blue-600 hover:text-blue-800 h-5 w-5 cursor-pointer flex-1
-                    }`}
-                      onClick={() => setActiveCategory(activeCategory - 1)}
-                    />
-                  ) : (
-                    <span className='h-5 w-5 flex-1' />
-                  )}
-
-                  {activeCategory < icons.length - 1 ? (
-                    <ChevronRightIcon
-                      className={`text-blue-600 hover:text-blue-800 h-5 w-5 cursor-pointer flex-1
-                    }`}
-                      onClick={() => setActiveCategory(activeCategory + 1)}
-                    />
-                  ) : (
-                    <span className='h-5 w-5 flex-1' />
-                  )}
-                </div>
-                <h2 className='font-semibold text-lg text-black dark:text-white'>
-                  {icons[activeCategory].title}
-                </h2>
-                <div className='flex w-12 justify-end'>
-                  <XMarkIcon
-                    className='text-blue-600 h-5 w-5 hover:text-blue-800'
-                    onClick={() => setActiveCategory(null)}
-                  />
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-        {/* start  */}
-        <div className='flex-col flex md:flex-row sm:justify-between h-full overflow-hidden'>
-          <div className='flex items-center justify-between sm:order-2 md:order-[initial] md:w-1/3 md:justify-center md:border-r-[0.3px] md:border-r-gray-300 dark:md:border-r-zinc-700 md:items-baseline'>
-            <div className='flex flex-row md:flex-col py-1 md:py-0 items-baseline gap-x-6 justify-evenly w-full md:mt-6 md:mx-4 md:bg-white dark:md:bg-zinc-800 md:rounded-xl md:pl-4'>
-              {icons.map((icon, index) => (
-                <button
-                  className={`grid place-content-center place-items-center gap-y-1.5 xs:w-[62px] cursor-pointer md:flex md:items-center md:justify-between group md:w-full md:flex-row-reverse md:px-6 md:py-3.5 md:pr-3 md:pl-0 ${
-                    index < icons.length - 1
-                      ? 'md:border-b md:border-[#3C3C43]/36 mdd:border-solid'
-                      : ''
-                  }`}
-                  key={icon.title}
-                  onClick={() =>
-                    activeCategory === index
-                      ? setActiveCategory(null)
-                      : setActiveCategory(index)
-                  }
-                >
-                  <span className='rounded-full bg-white dark:bg-zinc-700 group-hover:bg-stone-50 dark:group-hover:bg-zinc-600 grid place-items-center p-4 xs:w-[62px] xs:h-[62px] w-10 h-10 md:h-4 md:w-4 place-content-center'>
-                    <icon.icon
-                      className={`${
-                        activeCategory === index
-                          ? 'text-blue-600 dark:text-blue-400'
-                          : 'text-black dark:text-white'
-                      } h-5 w-5 stroke-2 group-hover:text-blue-400`}
-                    />
-                  </span>
-                  <span
-                    className={`${
-                      activeCategory === index ? 'text-blue-800 dark:text-blue-400' : 'text-black dark:text-white'
-                    } text-wrap text-center text-xs group-hover:text-blue-600 md:text-base`}
-                  >
-                    {icon.title}
-                  </span>
-                </button>
-              ))}
-            </div>
-          </div>
-          <hr className='border-[0.3px] border-gray-300 dark:border-zinc-700 border-solid sm:order-1 md:hidden' />
-          {activeCategory !== null ? (
-            <div className='flex flex-col py-5 sm:py-10 px-4 sm:px-12 gap-y-5 sm:gap-y-10 relative overflow-y-auto h-full w-full md:flex-1 bg-white/95 dark:bg-zinc-800/95'>
-              {activeCategory === 0 &&
-                skillCategories.map((category) => (
-                  <div key={category} className='flex flex-col gap-y-4'>
-                    <h2 className='text-sm font-semibold uppercase tracking-wide text-gray-500 dark:text-gray-400'>
-                      {category}
-                    </h2>
-                    <div className='flex flex-col gap-y-5 sm:gap-y-8'>
-                      {skills
-                        .filter((item) => item.category === category)
-                        .map((item) => (
-                          <div
-                            key={item.name}
-                            className='flex gap-x-5 sm:gap-x-8 w-full h-auto'
-                          >
-                            <div className='relative shrink-0'>
-                              {item.url ? (
-                                <a
-                                  href={
-                                    item.url.startsWith('http')
-                                      ? item.url
-                                      : `https://${item.url}`
-                                  }
-                                  target='_blank'
-                                  rel='noopener noreferrer'
-                                  className='block'
-                                >
-                                  <SkillLogo
-                                    name={item.name}
-                                    img={item.img}
-                                    iconKey={item.iconKey}
-                                    darkInvert={item.darkInvert}
-                                    iconColor={item.iconColor}
-                                  />
-                                </a>
-                              ) : (
-                                <SkillLogo
-                                  name={item.name}
-                                  img={item.img}
-                                  iconKey={item.iconKey}
-                                  darkInvert={item.darkInvert}
-                                  iconColor={item.iconColor}
-                                />
-                              )}
-                            </div>
-                            <div className='flex flex-col gap-y-2 min-w-0'>
-                              <h3 className='text-base font-semibold text-black dark:text-white'>
-                                {item.title}
-                              </h3>
-                              <div className='flex flex-col gap-y-1'>
-                                <p className='text-sm font-normal text-black dark:text-gray-300'>
-                                  {item.desc}
-                                </p>
-                                {item.url && (
-                                  <div className='block relative'>
-                                    <a
-                                      href={
-                                        item.url.startsWith('http')
-                                          ? item.url
-                                          : `https://${item.url}`
-                                      }
-                                      target='_blank'
-                                      rel='noopener noreferrer'
-                                      className='text-sm font-normal no-underline hover:underline text-violet-600 dark:text-violet-400 inline-flex items-center gap-x-2'
-                                    >
-                                      <LinkIcon height={16} />
-                                      {item.url.replace(/^https?:\/\//, '')}
-                                    </a>
-                                  </div>
-                                )}
-                              </div>
-                            </div>
-                          </div>
-                        ))}
-                    </div>
-                  </div>
-                ))}
-
-              {activeCategory === 1 &&
-                projects.map((item) => (
-                  <div key={item.title} className='flex gap-x-10 w-full h-auto'>
-                    <div className='relative sm:block hidden md:hidden lg:block'>
-                      {item.url ? (
-                        <a
-                          href={`https://${item.url}`}
-                          className='h-auto w-52 block'
-                        >
-                          <Image
-                            src={item.img}
-                            sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-                            className='rounded-lg shadow-xl'
-                            width={1}
-                            height={1}
-                            style={{
-                              height: 'auto',
-                              width: '100%'
-                            }}
-                            alt={item.title}
-                          />
-                        </a>
-                      ) : (
-                        <div className='h-auto w-24 block'>
-                          <Image
-                            src={item.img}
-                            sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-                            className='rounded-lg shadow-xl'
-                            width={1}
-                            height={1}
-                            style={{
-                              height: 'auto',
-                              width: '100%'
-                            }}
-                            alt={item.title}
-                          />
-                        </div>
-                      )}
-                    </div>
-                    <div className='flex flex-col gap-y-2'>
-                      <h3 className='text-base font-semibold text-black dark:text-white'>
-                        {item.title}
-                      </h3>
-                      <div className='relative sm:hidden block md:block lg:hidden'>
-                        {item.url ? (
-                          <a
-                            href={`https://${item.url}`}
-                            className='h-auto w-full block'
-                          >
-                            <Image
-                              src={item.img}
-                              sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-                              className='rounded-lg shadow-xl'
-                              width={1}
-                              height={1}
-                              style={{
-                                height: 'auto',
-                                width: '100%'
-                              }}
-                              alt={item.title}
-                            />
-                          </a>
-                        ) : (
-                          <div className='h-auto w-full block'>
-                            <Image
-                              src={item.img}
-                              sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-                              className='rounded-lg shadow-xl'
-                              width={1}
-                              height={1}
-                              style={{
-                                height: 'auto',
-                                width: '100%'
-                              }}
-                              alt={item.title}
-                            />
-                          </div>
-                        )}
-                      </div>
-                      <div className='flex flex-col gap-y-1'>
-                        <p className='text-sm font-normal text-black dark:text-gray-300'>
-                          {item.desc}
-                        </p>
-                        {item.url && (
-                          <div className='block relative'>
-                            <a
-                              href={`https://${item.url}`}
-                              className='text-sm font-normal no-underline hover:underline text-violet-600 dark:text-violet-400 inline-flex items-center gap-x-2'
-                            >
-                              <LinkIcon height={16} />
-                              {item.url}
-                            </a>
-                          </div>
-                        )}
-
-                        {item.github && (
-                          <div className='block relative'>
-                            <a
-                              href={item.github}
-                              className='text-sm font-normal no-underline hover:underline text-violet-600 dark:text-violet-400 inline-flex items-end gap-x-2'
-                            >
-                              <Image
-                                src='/images/icons/github.png'
-                                width={24}
-                                height={24}
-                                alt={`${item.title} GitHub`}
-                              />
-                              <span>GitHub</span>
-                            </a>
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  </div>
-                ))}
-
-              {activeCategory === 2 &&
-                workExperiences.map((item) => (
-                  <div
-                    key={`${item.company}-${item.startDate}`}
-                    className='flex gap-x-6 sm:gap-x-10 w-full h-auto'
-                  >
-                    <div className='relative shrink-0'>
-                      <CompanyLogo
-                        company={item.company}
-                        logo={item.logo}
-                        placeholder={item.logoPlaceholder}
-                      />
-                    </div>
-                    <div className='flex flex-col gap-y-2 min-w-0'>
-                      <div className='flex flex-col'>
-                        <h2 className='text-base font-semibold text-black dark:text-white'>
-                          {item.company}
-                        </h2>
-                        <h5 className='text-xs font-normal text-gray-600 dark:text-gray-400'>
-                          {item.startDate} –{' '}
-                          {item.isPresent ? 'Present' : item.endDate}
-                          {item.location ? ` · ${item.location}` : ''}
-                        </h5>
-                      </div>
-                      {item.projects.map((project) => (
-                        <div
-                          key={`${project.title}-${project.name}`}
-                          className='flex flex-col gap-y-1.5'
-                        >
-                          <div className='flex flex-col sm:flex-row sm:items-baseline sm:gap-x-2'>
-                            <h3 className='text-sm font-medium text-black dark:text-white'>
-                              {project.title}
-                            </h3>
-                            <span className='text-xs text-gray-500 dark:text-gray-400'>
-                              {project.name}
-                            </span>
-                          </div>
-                          <ul className='flex flex-col gap-y-1 list-disc pl-4'>
-                            {project.desc.map((desc: string) => (
-                              <li
-                                key={desc}
-                                className='text-sm font-normal text-gray-800 dark:text-gray-300'
-                              >
-                                {desc}
-                              </li>
-                            ))}
-                          </ul>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                ))}
-
-              {activeCategory === 3 &&
-                education.certificates.map((certificate) => (
-                  <div
-                    key={certificate.title}
-                    className='flex sm:gap-x-10 w-full h-auto'
-                  >
-                    <div className='relative sm:block md:hidden lg:block hidden'>
-                      <a
-                        href={`https://ude.my/${certificate.credentialID}`}
-                        className='h-auto w-52 block'
-                      >
-                        <Image
-                          src={`/images/education/${certificate.credentialID}.jpg`}
-                          sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-                          className='rounded-lg shadow-xl'
-                          width={1}
-                          height={1}
-                          style={{
-                            height: 'auto',
-                            width: '100%'
-                          }}
-                          alt={certificate.title}
-                        />
-                      </a>
-                    </div>
-                    <div className='flex flex-col gap-y-2'>
-                      <h3 className='text-base font-semibold text-black dark:text-white'>
-                        {certificate.title}
-                      </h3>
-                      <div className='relative block sm:hidden md:block lg:hidden'>
-                        <a
-                          href={`https://ude.my/${certificate.credentialID}`}
-                          className='h-auto w-full block'
-                        >
-                          <Image
-                            src={`/images/education/${certificate.credentialID}.jpg`}
-                            sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-                            className='rounded-lg shadow-xl'
-                            width={1}
-                            height={1}
-                            style={{
-                              height: 'auto',
-                              width: '100%'
-                            }}
-                            alt={certificate.title}
-                          />
-                        </a>
-                      </div>
-                      <div className='block relative'>
-                        <a
-                          href={`https://ude.my/${certificate.credentialID}`}
-                          className='text-sm font-normal no-underline hover:underline text-violet-600 dark:text-violet-400 inline-flex items-center gap-x-2'
-                        >
-                          <LinkIcon height={16} />
-                          {certificate.credentialID}
-                        </a>
-                      </div>
-                    </div>
-                  </div>
-                ))}
-            </div>
-          ) : (
-            <div className='flex items-center flex-col justify-center gap-y-2 flex-1 bg-white/95 dark:bg-zinc-800/95'>
-              <h2 className='text-base font-medium text-slate-800 dark:text-gray-200'>
-                Nothing to see here.
-              </h2>
-              <p className='text-xs font-normal text-slate-700 dark:text-gray-400'>
-                Please select one of the sidebar categories.
-              </p>
-            </div>
-          )}
-        </div>
-        {/* end */}
-      </div>
-    </main>
-  )
-}
-
-export default Portfolio
+export default PortfolioPage
